backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph>=1.0.0
langchain>=1.0.0
langchain-google-genai>=1.0.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic>=2.5.0
httpx>=0.24.0,<0.29.0
